srcs/billing-app/requirements.txt
Flask==2.0.3
psycopg2-binary==2.9.3
pika==1.2.0
